# evmrs: add feature alloc-reuse and reuse multiple instances for stack and memory