--- a/rust/src/types/memory.rs
+++ b/rust/src/types/memory.rs
@@ -1,15 +1,19 @@
-use std::{cmp::max, iter, sync::Mutex};
+#[cfg(feature = "alloc-reuse")]
+use std::sync::Mutex;
+use std::{cmp::max, iter};
 
 use crate::{
     types::{u256, FailStatus},
     utils::{word_size, Gas},
 };
 
+#[cfg(feature = "alloc-reuse")]
 static REUSABLE_MEMORY: Mutex<Vec<Vec<u8>>> = Mutex::new(Vec::new());
 
 #[derive(Debug)]
 pub struct Memory(Vec<u8>);
 
+#[cfg(feature = "alloc-reuse")]
 impl Drop for Memory {
     fn drop(&mut self) {
         let mut memory = Vec::new();
@@ -20,8 +24,12 @@ impl Drop for Memory {
 
 impl Memory {
     pub fn new(memory: &[u8]) -> Self {
+        #[cfg(not(feature = "alloc-reuse"))]
+        let mut m = Vec::new();
+        #[cfg(feature = "alloc-reuse")]
         let mut m = REUSABLE_MEMORY.lock().unwrap().pop().unwrap_or_default();
         m.clear();
+
         m.extend_from_slice(memory);
         Self(m)
     }